+++ changelog.rst
@@ -1,3 +1,13 @@
+1.0.0
+=======
+
+- First version
+- Fetch API, get pokemon list 
+- toggleViewMode (list or grid)
+- remember toggleViewMode
+- grid (4 col pc, 3 col tablet, 2 smartphone) 
+- responsive CSS
+
 0.0.0
 =======